Backend/requirements.txt
Flask~=2.2.3
flask_cors~=3.0.10
flask_sqlalchemy~=3.0.3
flask_session~=0.4.0
flask_socketio~=5.3.2
PyJWT~=2.6.0
redis~=4.5.1
psycopg2-binary==2.9.5
bcrypt~=4.0.1
python-dotenv~=1.0.0
